# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Cupreous Dating App"
    },
    "host": "http://54.152.239.30/:9118",
    "basePath": "/",
    "paths": {
...
        "/payment/createCustomerSubscription":{
            "post": {
                "tags": ["PaymentGateWay"],
                "summary": "To Create Subscription",
                "description": "This for Creating Subscription",
                "operationId": "Subscription",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "like",
                        "in": "body",
                        "description": "To pass a like from one user To another User",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Create_Subscription_Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid email/password supplied"
                    },
                    "406": {
                        "description": "Something Went Wrong"
                    }
                }
            }
        },
        "/payment/createProduct":{
            "post": {
                "tags": ["PaymentGateWay"],
                "summary": "To Create Product",
                "description": "This to create Product",
                "operationId": "createProduct",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "createProduct",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Create_Product_Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Create_Product_Response"
                        }
                    },
                    "400": {
                        "description": "Invalid email/password supplied"
                    },
                    "406": {
                        "description": "Something Went Wrong"
                    }
                }
            }
        },
...
    },
...
    "definitions": {
...
        "Create_Subscription_Request":{
            "type":"object",
            "properties":{
                "customerEmail":{
                    "type":"string",
                    "required":true
                },
                "stripeToken": {
                    "type":"string",
                    "required":true
                },
                "planId":{
                    "type":"string",
                    "required":true
                }
            }
        },
...
        "Create_Product_Request":{
            "type":"object",
            "properties":{
                "name":{
                    "type":"string"
                },
                "type":{
                    "type":"string"
                }
            }
        },
        "Create_Product_Response":{
            "type":"object",
            "properties":{
                "id":{
                    "type":"string"
                },
                "active":{
                    "type" : false
                },
                "name":{
                    "type":"string"
                },
                "metadata":{
                    "type":"object"
                }
            }
        },
...
    },
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"]
}